+++ account_customer_wallet/tests/__init__.py
@@ -1,1 +1,2 @@
+from . import common
 from . import test_balance
